Store snapshot's stake modifier inside the snapshot

+++ src/test/snapshot_iterator_tests.cpp
@@ -19,7 +19,7 @@ BOOST_AUTO_TEST_CASE(snapshot_iterator) {
   {
     // generate the snapshot
     uint256 blockHash = uint256S("aa");
-    snapshot::Indexer idx(0, uint256(), blockHash, 3, 2);
+    snapshot::Indexer idx(0, uint256(), blockHash, uint256(), 3, 2);
     for (uint32_t i = 0; i < msgsToGenerate; ++i) {
       snapshot::UTXOSubset subset;
       subset.m_txId.SetHex(std::to_string(i));
